+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.12.5](https://github.com/witchesofthehill/manabrew/compare/v1.12.4...v1.12.5) (2026-07-20)
+
+## [1.12.4](https://github.com/witchesofthehill/manabrew/compare/v1.12.3...v1.12.4) (2026-07-19)
+
+### Fixes
+
+* x cost prompting, prompt displays, grid ([#509](https://github.com/witchesofthehill/manabrew/issues/509)) ([395a943](https://github.com/witchesofthehill/manabrew/commit/395a9439dc7496d2e32a5db024cf9676ac66e253))
+
+## [1.12.3](https://github.com/witchesofthehill/manabrew/compare/v1.12.2...v1.12.3) (2026-07-19)
+
+### Fixes
+
+* **forge:** offer phyrexian life payment in action space ([#462](https://github.com/witchesofthehill/manabrew/issues/462)) ([ad76eef](https://github.com/witchesofthehill/manabrew/commit/ad76eefcb8d8344334650f71f295e9f82c3653d0))
+
 ## [1.12.2](https://github.com/witchesofthehill/manabrew/compare/v1.12.1...v1.12.2) (2026-07-19)
 
 ## [1.12.1](https://github.com/witchesofthehill/manabrew/compare/v1.12.0...v1.12.1) (2026-07-17)